gestoppes/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = "^3.11"
llama-agents = "^0.0.3"
llama-index-embeddings-openai = "^0.1.10"
llama-index-llms-openai = "^0.1.23"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "0.2.6"

[tool.poetry.dependencies.llama-index-readers-database]
version = "^0.1.3"

[tool.poetry.dependencies.pymysql]
version = "^1.1.0"
extras = [ "rsa" ]

[tool.poetry.dependencies.psycopg2]
version = "^2.9.9"

[tool.poetry.dependencies.llama-index-indices-managed-llama-cloud]
version = "^0.2.5"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"